page/workout | fix layout
