+++ craco.config.js
@@ -22,7 +22,6 @@ module.exports = {
           '@api': './src/api',
           '@utils': './src/utils',
           '@redux': './src/redux',
-          '@routers': './src/routers',
         },
       },
     },
