+++ src/components/header/header.tsx
@@ -1,0 +1,6 @@
+import * as Styled from './styled.header';
+
+
+export const Header = () => (
+  <Styled.HeaderWrapper />
+);
