+++ src/components/header/styled.header.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+
+export const HeaderWrapper = styled.div`
+    display: flex;
+    width: 100%;
+`;
