+++ src/layouts/layout/styled.layout.tsx
@@ -2,6 +2,9 @@ import styled from 'styled-components';
 
 
 export const LayoutWrapper = styled.div`
+    display: flex;
+    flex-direction: column;
     width: 100%;
     height: 100%;
+    padding: 0 140px;
 `;
